+++ library/core/src/main/java/com/google/android/exoplayer2/ExoPlayer.java
@@ -185,6 +185,10 @@ public interface ExoPlayer extends Player {
    */
   Looper getPlaybackLooper();
 
+  @Override
+  @Nullable
+  ExoPlaybackException getPlaybackError();
+
   /**
    * Prepares the player to play the provided {@link MediaSource}. Equivalent to
    * {@code prepare(mediaSource, true, true)}.