+++ testutils_robolectric/src/main/java/com/google/android/exoplayer2/testutil/StubExoPlayer.java
@@ -17,6 +17,7 @@ package com.google.android.exoplayer2.testutil;
 
 import android.os.Looper;
 import android.support.annotation.Nullable;
+import com.google.android.exoplayer2.ExoPlaybackException;
 import com.google.android.exoplayer2.ExoPlayer;
 import com.google.android.exoplayer2.PlaybackParameters;
 import com.google.android.exoplayer2.Player;
@@ -60,6 +61,11 @@ public abstract class StubExoPlayer implements ExoPlayer {
 
   @Override
   public int getPlaybackState() {
+    throw new UnsupportedOperationException();
+  }
+
+  @Override
+  public ExoPlaybackException getPlaybackError() {
     throw new UnsupportedOperationException();
   }